+++ modules/fabric_1.20.1/src/main/java/com/armilp/ifreq/network/ModPackets.java
@@ -18,5 +18,9 @@ public class ModPackets {
                     WalkiePowerPacket.TYPE,
                     WalkiePowerPacket::receive
             );
+            ServerPlayNetworking.registerGlobalReceiver(
+                    WalkieVolumePacket.TYPE,
+                    WalkieVolumePacket::receive
+            );
         }
 }