+++ modules/fabric_1.20.1/changelog.md
@@ -2,7 +2,8 @@
 
 ### Added
   - 3d model for the walkie-talkie item
-  - walkie-talkie gui can be open while in the blet slot of the Accessories mod hitting R key
+  - walkie-talkie gui can be open while in the blet slot of the Accessories mod by hitting R key
   - Config for editing range of the walkie-talkie
+  - Added Beep sound slider to walkie-talkie gui in case it too loud.
 
 This is a beta version, so expect some bugs and missing features.